Add h2 integration test and change PK from Int to Long due to hibernate validation error

+++ src/main/java/com/avimedical/appointments/scheduling/adapters/database/AppointmentCrudRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-interface AppointmentCrudRepository extends CrudRepository<AppointmentEntity, Long> {
+interface AppointmentCrudRepository extends CrudRepository<AppointmentEntity, Integer> {
 }
